Trait hydra_llvm​::codegen​::StmtGenerator​

pub trait StmtGenerator {
    fn gen_code(&self, &mut Builder);
    fn register_state_vars(&self, &mut GenBuilder) -> uint;
    fn gen_gen_code(&self, &mut GenGenState, &mut Builder);
}

Required Methods

fn gen_code(&self, &mut Builder)

fn register_state_vars(&self, &mut GenBuilder) -> uint

Add declared variable types to vec so that a generator state object can be created...push the number of vars u need in so that you know where to start from on your save restore

fn gen_gen_code(&self, &mut GenGenState, &mut Builder)

Generate your regular code but also break to the end of the save and restore blocks provided to save and restore state variables so this can be done in one pass